add some header files
add our version of klee.h. This is stripped version of klee.h with
declaration of just the functions that we need. It does not contain any
nested #include, so cannot break build on different environment.

add symbiotic.h. This file contains declarations that we need
(currently just __assert_fail)