insert __VERIFIER_error() twice
I know: what? This is kind of workaround. When in the code is

ERROR:
call __VERIFIER_error(...)
br %ERROR

then the br causes the call to be sliced away. If we put there
two calls, one will be there and everything works well. This
should be fixed in slicer, but until then this patch will take
care of it